Seperated interrupts and exceptions. Changed segment registers to 16 bit

+++ src/kernel/boot/idt/interrupt_handler.h
@@ -40,10 +40,10 @@ namespace kernel::boot::idt {
     };
 
     struct SegmentRegisters {
-        uint32_t gs;
-        uint32_t fs;
-        uint32_t es;
-        uint32_t ds;
+        uint16_t gs;
+        uint16_t fs;
+        uint16_t es;
+        uint16_t ds;
     };
 
     struct InterruptInfo {
